simple_bot/requirements.txt
botframework-connector>=4.0.0.a2
tornado==3.2.1